Cambios pequeños y mejoras para turno. No esta operativo todavia, le falta orden y mas construccion.

+++ Controladores/TurnoControlador.cs
@@ -38,7 +38,7 @@ namespace Controlador
             return TurnoData.SearchByParameters(searchText);
         }
 
-        public static TurnoView GetByPrimaryKeyView(int idTurno)
+        public static TurnoView GetByPrimaryKeyView(long idTurno)
         {
             return TurnoData.GetByPrimaryKeyView(idTurno);
         }
